Move lib/dnssd/minimal_mdns.records -> lib/dnssd/wire
