+++ src/BUILD.gn
@@ -210,7 +210,7 @@ if (chip_build_tests) {
     }
 
     if (current_os != "zephyr") {
-      tests += [ "${chip_root}/src/lib/dnssd/minimal_mdns/records/tests" ]
+      tests += [ "${chip_root}/src/lib/dnssd/wire/records/tests" ]
     }
 
     if (current_os != "zephyr") {
