+++ src/lib/dnssd/BUILD.gn
@@ -104,7 +104,7 @@ static_library("dnssd") {
     ]
     public_deps += [
       "${chip_root}/src/lib/dnssd/minimal_mdns",
-      "${chip_root}/src/lib/dnssd/minimal_mdns/records",
+      "${chip_root}/src/lib/dnssd/wire/records",
       "${chip_root}/src/lib/dnssd/wire",
       "${chip_root}/src/tracing",
       "${chip_root}/src/tracing:macros",
