+++ src/lib/dnssd/minimal_mdns/responders/Srv.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include <lib/dnssd/minimal_mdns/records/Srv.h>
+#include <lib/dnssd/wire/records/Srv.h>
 #include <lib/dnssd/minimal_mdns/responders/RecordResponder.h>
 
 namespace mdns {
