+++ src/lib/dnssd/minimal_mdns/tests/TestAdvertiser.cpp
@@ -27,9 +27,9 @@
 #include <lib/dnssd/MinimalMdnsServer.h>
 #include <lib/dnssd/minimal_mdns/Query.h>
 #include <lib/dnssd/minimal_mdns/QueryBuilder.h>
-#include <lib/dnssd/minimal_mdns/records/Ptr.h>
-#include <lib/dnssd/minimal_mdns/records/Srv.h>
-#include <lib/dnssd/minimal_mdns/records/Txt.h>
+#include <lib/dnssd/wire/records/Ptr.h>
+#include <lib/dnssd/wire/records/Srv.h>
+#include <lib/dnssd/wire/records/Txt.h>
 #include <lib/dnssd/minimal_mdns/tests/CheckOnlyServer.h>
 #include <lib/dnssd/wire/QName.h>
 #include <lib/support/tests/ExtraPwTestMacros.h>
