+++ src/lib/dnssd/tests/BUILD.gn
@@ -79,7 +79,7 @@ if (pw_enable_fuzz_test_targets) {
       public_deps = [
         "${chip_root}/src/lib/dnssd",
         "${chip_root}/src/lib/dnssd/minimal_mdns",
-        "${chip_root}/src/lib/dnssd/minimal_mdns/records",
+        "${chip_root}/src/lib/dnssd/wire/records",
         "${chip_root}/src/lib/dnssd/wire",
         "${chip_root}/src/platform/logging:default",
       ]
