+++ src/lib/dnssd/tests/FuzzMdnsResolverPW.cpp
@@ -58,11 +58,11 @@
 #include <lib/dnssd/Types.h>
 #include <lib/dnssd/minimal_mdns/Parser.h>
 #include <lib/dnssd/minimal_mdns/RecordData.h>
-#include <lib/dnssd/minimal_mdns/records/IP.h>
-#include <lib/dnssd/minimal_mdns/records/Ptr.h>
-#include <lib/dnssd/minimal_mdns/records/ResourceRecord.h>
-#include <lib/dnssd/minimal_mdns/records/Srv.h>
-#include <lib/dnssd/minimal_mdns/records/Txt.h>
+#include <lib/dnssd/wire/records/IP.h>
+#include <lib/dnssd/wire/records/Ptr.h>
+#include <lib/dnssd/wire/records/ResourceRecord.h>
+#include <lib/dnssd/wire/records/Srv.h>
+#include <lib/dnssd/wire/records/Txt.h>
 #include <lib/dnssd/wire/QName.h>
 #include <lib/dnssd/wire/RecordWriter.h>
 #include <lib/support/CHIPMem.h>
