+++ src/lib/dnssd/tests/TestIncrementalResolve.cpp
@@ -24,11 +24,11 @@
 
 #include <lib/core/StringBuilderAdapters.h>
 #include <lib/dnssd/ServiceNaming.h>
-#include <lib/dnssd/minimal_mdns/records/IP.h>
-#include <lib/dnssd/minimal_mdns/records/Ptr.h>
-#include <lib/dnssd/minimal_mdns/records/ResourceRecord.h>
-#include <lib/dnssd/minimal_mdns/records/Srv.h>
-#include <lib/dnssd/minimal_mdns/records/Txt.h>
+#include <lib/dnssd/wire/records/IP.h>
+#include <lib/dnssd/wire/records/Ptr.h>
+#include <lib/dnssd/wire/records/ResourceRecord.h>
+#include <lib/dnssd/wire/records/Srv.h>
+#include <lib/dnssd/wire/records/Txt.h>
 #include <lib/dnssd/wire/tests/QNameStrings.h>
 #include <lib/support/ScopedMemoryBuffer.h>
