+++ src/lib/dnssd/wire/records/BUILD.gn
@@ -29,7 +29,6 @@ static_library("records") {
     "${chip_root}/src/inet",
     "${chip_root}/src/lib/core",
     "${chip_root}/src/lib/dnssd/wire",
-    "${chip_root}/src/platform",
   ]
 
   cflags = [ "-Wconversion" ]
