+++ src/lib/dnssd/wire/records/Srv.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include <lib/dnssd/minimal_mdns/records/ResourceRecord.h>
+#include <lib/dnssd/wire/records/ResourceRecord.h>
 
 namespace mdns {
 namespace Minimal {
