+++ src/lib/dnssd/wire/records/tests/BUILD.gn
@@ -33,6 +33,6 @@ chip_test_suite("tests") {
   public_deps = [
     "${chip_root}/src/lib/core",
     "${chip_root}/src/lib/core:string-builder-adapters",
-    "${chip_root}/src/lib/dnssd/minimal_mdns/records",
+    "${chip_root}/src/lib/dnssd/wire/records",
   ]
 }
